+++ aquadoggo/src/graphql/mod.rs
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
 pub mod client;
+pub mod pagination;
 pub mod replication;
-pub mod response;
 pub mod scalars;
 mod schema;
